Layer panel refactor
